Control drawer state
