+++ projectBlueWater/src/main/java/com/lasthopesoftware/bluewater/client/playback/nowplaying/view/activity/viewmodels/ControlDrawerState.kt
@@ -1,0 +1,6 @@
+package com.lasthopesoftware.bluewater.client.playback.nowplaying.view.activity.viewmodels
+
+interface ControlDrawerState {
+	fun showDrawer()
+	fun hideDrawer()
+}
