+++ projectBlueWater/src/main/res/layout/control_now_playing.xml
@@ -5,7 +5,7 @@
 
     <data>
         <import type="android.view.View"/>
-        <variable name="vm" type="com.lasthopesoftware.bluewater.client.playback.nowplaying.view.activity.viewmodels.NowPlayingViewModel"/>
+        <variable name="vm" type="com.lasthopesoftware.bluewater.client.playback.nowplaying.view.activity.viewmodels.NowPlayingFilePropertiesViewModel"/>
     </data>
 
     <androidx.coordinatorlayout.widget.CoordinatorLayout
